use prepend; include did not work outside of rails console for some reason
